Fixing tcp transport issues

+++ components/event-processor-manager/org.wso2.carbon.event.processor.manager.commons/src/main/java/org/wso2/carbon/event/processor/manager/commons/transport/client/TCPEventPublisher.java
@@ -128,8 +128,8 @@ public class TCPEventPublisher {
                     buf.putDouble((Double) event[i]);
                     continue;
                 case STRING:
-                    short length = (short) ((String) event[i]).length();
-                    buf.putShort(length);
+                    int length = ((String) event[i]).length();
+                    buf.putInt(length);
                     stringDataIndex[stringIndex] = i;
                     stringIndex++;
                     stringSize += length;
